Move modal to AichatView
