+++ apps/src/aichat/views/ChatWorkspace.tsx
@@ -4,12 +4,8 @@ import Tabs, {TabsProps} from '@code-dot-org/component-library/tabs';
 import React, {useCallback, useEffect, useMemo, useState} from 'react';
 import {useSelector} from 'react-redux';
 
-import TeacherOnboardingModal from '@cdo/apps/aichat/views/TeacherOnboardingModal';
-import ChatWarningModal from '@cdo/apps/aiComponentLibrary/warningModal/ChatWarningModal';
 import {useAppDispatch, useAppSelector} from '@cdo/apps/util/reduxHooks';
-import {tryGetLocalStorage, trySetLocalStorage} from '@cdo/apps/utils';
 
-import {ModalTypes} from '../constants';
 import aichatI18n from '../locale';
 import {
   clearChatMessages,
@@ -17,7 +13,6 @@ import {
   fetchUserChatHistory,
   selectAllVisibleMessages,
   selectMultimodalEnabled,
-  setShowModalType,
 } from '../redux';
 import {ChatButton} from '../types';
 import {getShortName} from '../utils';
@@ -56,11 +51,8 @@ const ChatWorkspace: React.FunctionComponent<ChatWorkspaceProps> = ({
   const [selectedTab, setSelectedTab] =
     useState<WorkspaceTeacherViewTab | null>(null);
 
-  const {showModalType, studentChatHistory} = useAppSelector(
-    state => state.aichat
-  );
+  const {studentChatHistory} = useAppSelector(state => state.aichat);
   const currentLevelId = useAppSelector(state => state.progress.currentLevelId);
-  const isUserTeacher = useAppSelector(state => state.currentUser.isTeacher);
   const visibleItems = useSelector(selectAllVisibleMessages);
   const currentUserId = useAppSelector(state => state.currentUser.userId);
 
@@ -101,25 +93,6 @@ const ChatWorkspace: React.FunctionComponent<ChatWorkspaceProps> = ({
     () => selectedTab !== WorkspaceTeacherViewTab.STUDENT_CHAT_HISTORY,
     [selectedTab]
   );
-
-  useEffect(() => {
-    const modalToShow = () => {
-      if (!isUserTeacher) {
-        return ModalTypes.WARNING;
-      }
-
-      const teacherSawAichatOnboardingModal = tryGetLocalStorage(
-        'teacherSawAichatOnboarding',
-        'no'
-      );
-
-      return teacherSawAichatOnboardingModal === 'yes'
-        ? undefined
-        : ModalTypes.TEACHER_ONBOARDING;
-    };
-
-    dispatch(setShowModalType(modalToShow()));
-  }, [isUserTeacher, dispatch]);
 
   useEffect(() => {
     // If we are viewing as a student, default to the student chat history tab if tab is not yet selected.
@@ -177,35 +150,10 @@ const ChatWorkspace: React.FunctionComponent<ChatWorkspaceProps> = ({
     tabPanelsContainerClassName: moduleStyles.tabPanelsContainer,
   };
 
-  const ChatModal = useMemo(
-    () =>
-      showModalType === ModalTypes.TEACHER_ONBOARDING
-        ? TeacherOnboardingModal
-        : showModalType === ModalTypes.WARNING
-        ? ChatWarningModal
-        : undefined,
-    [showModalType]
-  );
-
-  const onCloseModal = useCallback(() => {
-    // We only want to show the teacher onboarding modal the first time a teacher user
-    // interacts with the aichat tool. Thus, we store a value in local storage when
-    // closing the modal.
-    if (
-      isUserTeacher &&
-      showModalType === ModalTypes.TEACHER_ONBOARDING &&
-      tryGetLocalStorage('teacherSawAichatOnboarding', 'no') !== 'yes'
-    ) {
-      trySetLocalStorage('teacherSawAichatOnboarding', 'yes');
-    }
-    dispatch(setShowModalType(undefined));
-  }, [dispatch, isUserTeacher, showModalType]);
-
   const multimodalEnabled = useAppSelector(selectMultimodalEnabled);
 
   return (
     <div id="chat-workspace-area" className={moduleStyles.chatWorkspace}>
-      {ChatModal && <ChatModal onClose={onCloseModal} />}
       {selectedStudent ? (
         <Tabs {...tabArgs} />
       ) : (
